pyproject.toml
[project]
name = "jellys"
version = "0.1.0"
description = "2D differentiable MPM simulation for jellyfish-inspired soft robot co-design"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "taichi>=1.7.0; platform_machine == 'x86_64'",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[project.scripts]
jellys = "jellys.main:main"

[tool.uv]
package = true
